requirements.txt
# VALD Reference Data Upload Project Dependencies
# Core data processing libraries
pandas>=1.5.0
numpy>=1.21.0

# Google Cloud Platform libraries
google-cloud-bigquery>=3.0.0
google-oauth2>=4.0.0
pandas-gbq>=0.17.0

# HTTP and API libraries
requests>=2.28.0
aiohttp>=3.8.0

# Environment and configuration
python-dotenv>=0.19.0

# Note: asyncio is part of Python's standard library (Python 3.7+)
# Note: uuid, datetime, json, os, sys are part of Python's standard library
